# Improved for loop for faster negotiate response