Started auditing SignalR's `for`/`foreach` loops to look for obvious perf gains and found this one immediately.

Before:

Method |     Mean |    Error |   StdDev |   Median |        Op/s |  Gen 0 | Allocated |
--------------------------------- |---------:|---------:|---------:|---------:|------------:|-------:|----------:|
WriteResponse_MemoryBufferWriter | 595.1 ns | 14.50 ns | 33.90 ns | 583.1 ns | 1,680,415.1 | 0.0048 |     144 B |

After:

Method |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
--------------------------------- |---------:|---------:|---------:|------------:|-------:|----------:|
WriteResponse_MemoryBufferWriter | 503.4 ns | 9.909 ns | 14.21 ns | 1,986,464.1 | 0.0019 |      64 B |