@BrennanConroy 2019-07-02
I'll check, but I'm 90% sure it is since I merged it early in the morning